pyproject.toml
[tool.poetry]
name = "chesseval"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chess = "^1.10.0"
openai = "^1.19.0"
stockfish = "^3.28.0"
anthropic = "^0.25.2"
google-generativeai = "^0.5.0"
fireworks-ai = "^0.13.0"
python-dotenv = "^1.0.1"
ruff = "^0.3.7"
pre-commit = "^3.7.0"
mypy = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
